requirements.txt
Django==5.0.1 
requests
python-dotenv
openai
fatsecret
